fix: wrong data display
While the logic is correct, the way of the data presented previously was
wrong since there's no guarantee that every `jenis` of `Penelitian`
would exist every year. This makes the ordering wrong because of the
difference of the number of year which as `penelitian` on them.

This change may be applied to other object as well, but not in the near
time